Tried to get more sensible printing of floating point numbers
Results are quite disappointing. Resolution on long doubles is nothing like as good as I hoped; they're out by one part in 10^20.

All unit tests except one pass, and the one that doesn't is a very minor rounding issue, so I'm calling it good.

+++ src/integer.c
@@ -21,7 +21,7 @@
  * but only integers and reals are so far implemented.
  */
 long double numeric_value( struct cons_pointer pointer ) {
-    double result = NAN;
+    long double result = NAN;
     struct cons_space_object *cell = &pointer2cell( pointer );
 
     if ( integerp( pointer ) ) {
